Changed the XCOMM_SetDacSamplingRate() parameter and return type.

+++ fmcomms1/Common/xcomm.c
@@ -250,7 +250,7 @@ int32_t XCOMM_InitTx(XCOMM_DefaultInit* pDefInit)
         return -1;
 
     /* Set the AD9122 sampling rate */
-    if(XCOMM_SetDacSamplingRate((uint32_t)pDefInit->dacSamplingRate) < 0)
+    if(XCOMM_SetDacSamplingRate(pDefInit->dacSamplingRate) < 0)
         return -1;
 
     return 0;
@@ -1003,7 +1003,7 @@ int32_t XCOMM_SetDacCenterShiftFreq(int32_t shift)
 * @return If success, return exact calculated rate in Hz
 *         if error, return -1
 ******************************************************************************/
-int32_t XCOMM_SetDacSamplingRate(uint32_t rate)
+int64_t XCOMM_SetDacSamplingRate(uint64_t rate)
 {
     int32_t sampleRate;
